Refactor, move models to submodule and export in __init__
